Trimmed Proposals/Create
The create proposal page was just a lot of extra tags that we never even
arrived to, since when you clicked create, you were moved to the
proposal update page anyways. I deleted most of it. Create now makes an
empty proposal, and update allows the user to go back and update the
proposal as necessary